+++ hive_ce_inspector/lib/model/hive_internal.dart
