+++ hive_ce_inspector/lib/service/connect_client.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:hive_ce_inspector/hive_internal.dart';
+import 'package:hive_ce_inspector/model/hive_internal.dart';
 import 'package:vm_service/vm_service.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
