+++ hive_ce_inspector/lib/widget/connected_layout.dart
@@ -1,8 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:hive_ce_inspector/connect_client.dart';
-import 'package:hive_ce_inspector/hive_internal.dart';
+import 'package:hive_ce_inspector/service/connect_client.dart';
+import 'package:hive_ce_inspector/model/hive_internal.dart';
+import 'package:hive_ce_inspector/model/box_data.dart';
 
 class ConnectedLayout extends StatefulWidget {
   final ConnectClient client;
@@ -12,21 +13,6 @@ class ConnectedLayout extends StatefulWidget {
 
   @override
   State<StatefulWidget> createState() => _ConnectedLayoutState();
-}
-
-class BoxData {
-  final String name;
-  final Map<Object, InspectorFrame> frames;
-  final bool open;
-
-  BoxData({
-    required this.name,
-    Map<Object, InspectorFrame>? frames,
-    this.open = true,
-  }) : frames = frames ?? {};
-
-  BoxData copyWith({bool? open}) =>
-      BoxData(name: name, frames: frames, open: open ?? this.open);
 }
 
 class _ConnectedLayoutState extends State<ConnectedLayout> {
