+++ hive_ce_inspector/lib/widget/error_screen.dart
